Moved sysv_patcher function into wrappers
